splitting tests. still need refactor.
