+++ test/unit/math/prim/scal/prob/bernoulli_test.cpp
@@ -6,68 +6,37 @@
 TEST(ProbDistributionsBernoulli, error_check) {
   boost::random::mt19937 rng;
   EXPECT_NO_THROW(stan::math::bernoulli_rng(0.6,rng));
-  EXPECT_NO_THROW(stan::math::bernoulli_logit_rng(-3.5,rng));
 
   EXPECT_THROW(stan::math::bernoulli_rng(1.6,rng),std::domain_error);
   EXPECT_THROW(stan::math::bernoulli_rng(-0.6,rng),std::domain_error);
   EXPECT_THROW(stan::math::bernoulli_rng(stan::math::positive_infinity(),rng),
                std::domain_error);
-  EXPECT_THROW(stan::math::bernoulli_logit_rng(stan::math::positive_infinity(),rng),
-               std::domain_error);
-}
-
-/*
- * Uses a chi-squared test to check that counts are reasonably consistent
- * with draws from a Bernoulli distribution..
- */
-void assert_bernoulli(const std::vector<int>& counts, const std::vector<double>& expected_counts, double tolerance) {
-  boost::math::chi_squared dist(1);
-  
-  EXPECT_EQ(counts.size(), 2);
-  EXPECT_EQ(expected_counts.size(), 2);
-  
-  double chi = 0;
-  for (int i=0; i<2; ++i) {
-    const double discrepancy = expected_counts[i] - counts[i];
-    chi += discrepancy * discrepancy / expected_counts[i];
-  }
-  
-  const double chi_threshold = quantile(complement(dist, tolerance));
-  EXPECT_TRUE(chi < chi_threshold);
 }
 
 TEST(ProbDistributionsBernoulli, chiSquareGoodnessFitTest) {
   boost::random::mt19937 rng;
   int N = 10000;
-
-  const int tmp1[2] = {0, 0};
-  std::vector<int> counts(tmp1, tmp1 + 2);
-
-  const double tmp2[2] = {N * (1 - 0.4), N * 0.4};
-  const std::vector<double> expected_counts(tmp2, tmp2 + 2);
-
-  for (int i=0; i<N; ++i) {
-    ++counts[stan::math::bernoulli_rng(0.4, rng)];
-  }
-  assert_bernoulli(counts, expected_counts, 1e-6);
-}
-
-
-TEST(ProbDistributionsBernoulli, logitChiSquareGoodnessFitTest) {
-  boost::random::mt19937 rng;
-  int N = 10000; // number of samples
+  boost::math::bernoulli_distribution<>dist (0.4);
+  boost::math::chi_squared mydist(1);
  
-  double parameter = -0.5; // logit-transformed probability
-  double prob = stan::math::inv_logit(-0.5); // actual probability
+  int bin[2] = {0, 0};
+  double expect [2] = {N * (1 - 0.4), N * (0.4)};
 
-  const int tmp1[2] = {0, 0};
-  std::vector<int> counts(tmp1, tmp1 + 2);
+  int count = 0;
 
-  const double tmp2[2] = {N * (1 - prob), N * prob};
-  const std::vector<double> expected_counts(tmp2, tmp2 + 2);
+  while (count < N) {
+    int a = stan::math::bernoulli_rng(0.4,rng);
+    if(a == 1)
+      ++bin[1];
+    else
+      ++bin[0];
+    count++;
+   }
 
-  for (int i=0; i<N; ++i) {
-    ++counts[stan::math::bernoulli_logit_rng(parameter, rng)];
-  }
-  assert_bernoulli(counts, expected_counts, 1e-6);
+  double chi = 0;
+
+  for(int j = 0; j < 2; j++)
+    chi += ((bin[j] - expect[j]) * (bin[j] - expect[j]) / expect[j]);
+
+  EXPECT_TRUE(chi < quantile(complement(mydist, 1e-6)));
 }
